+++ CHANGELOG.md
@@ -8,10 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming release template] - 1971-11-03
 
 ### Added
--
+- Add Form/Tooltip component.
 
 ### Changed
--
+- InputField, Radio, Checkbox, Dropdown, Textarea components.
 
 ### Deprecated
 -